facial features component/requirements.txt
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
python-dotenv>=0.19.0

